fixes concurrent iterate and write with current height iterator
This situation came up in a busy node, staked for chains 0001,0003,0004,0005,0009,0010,0021,0022,0023,0024,0025,0026,0027,0028 simultaneously. We need to observe the effects of this change in the CPU and general performance: if it's too big a dip in performance, we may have to consider the more complex alternative of more closely mimicking the iavl's iterator generation technique.