partial implementation of the fast archival store
Introducing most of the fast archival srore implementation -- crucially missing the code needed to make it usable from
the multistore, the "clean-slate" process needed for liveState, and the pruning for the iavl store whose merkle tree
we're depending on.

The store consists of an iavl (used solely for its merkle tree, hence named merkle in the struct), a state db to store
the latest information before it is committed (called liveState), and the archival store, which is a key value store
with all the information in the history of the chain.

Another change done here is the change of a method's signature from an interface to the actual object; it's not counting
on the interface in any case.