partial implementation of the fast archival store
fast archival store is what previously was used in the "waves" implementation.

It consists of an iavl (used solely for its merkle tree, hence named merkle in the struct), a state db to store the latest information before it is committed (called liveState), and the archival store, which is a key value store with all the information in the history of the chain.