inlining start function, extracting signal capture
the anonymous function assigned to startCmd's Run was acting as a passthrough for the start function, so I inlined it.

The signal capturing code, on the other hand, was doing a lot of work that wasn't related to the startup sequence - basically waiting for the end signals from the OS. I extracted the behavior into its own function.