introducing hidden fastarchival flag in start cmd
In order to implement fastarchival we're making the CommitMultiStore abstract over the actual storage mechanism, which wasn't the case before. So the flag is introduced but not wired. Changing the code around so that the CommitMultiStore handles the persistence of the data while abstracting over the details will imply moving the behavior of the app.OpenApplicationDB function elsewhere, which can be delicate since that is an exported function.

There are a few warnings in this commit that will not be addressed, all in app/config.go. I will not hastily try to fix those at the time since it will veer into configuration loading, which will likely be touched upon later when the persistence replacement effort nears a close.