Remixing multistore configuration code path
In order to get the information needed for the multistore to manage its storage, the stack of functions called from the start command up to the place the multistore is configured is being remixed.

Makes the genesisType get computed outside of the start function for single-responsibility purposes.
Marks keybase as deprecated in the start command and removes it from the call stack as it's ultimately unused.
AppCreator function was removed from the client initialization.
Removes the `config` and `Config` structs from app package.
Removes aliases from the tendermint/tendermint/config package, as they are no longer needed to avoid shadowing.