Addressing #1501 by making set- and get-validator LeanPocket aware
Several things are happening here under that title; for one, get-validator and set-validator now detect whether
LeanPocket is on; if it is, then they output a message suggesting you might be trying to run the plural version of the
commands and exits.

The messages in set-validators and get-validators were similarly adapted to suggest you might be trying to run the
singular version of the command before exiting.

Last and least, a the setValidatorsLean command was slightly edited so the whitespace around the check for LeanPocket
matches the others.