making return values unnamed
I am pretty uncomfortable about the semantics of this code; I end up declaring a ton of `err` variables for no reason other than they are the second parameter, and the first one needs to be created.

The tendency towards being terse is fighting the common code structuring patterns and the semantics in a royal rumble where the winner gets to still not be satisfactory, so I'll just keep this in a recognizable style and  call it quits until a guiding principle can show up and help us never face such mixed solutions of sometimes creating and sometimes reusing a variable.