Clear the global session cache when a node is unjailed (#1529)
This is a fix for a consensus failure with `wrong Block.Header.AppHash`,
which is probably the same issue as #1478.

### Issue description:

For a node to be selected for a session, it must be valid not only at
the start of a session but also at the end of a session. If a node is
edit-staked, jailed, or unjailed in the middle of a session, a servicer
set of the session is changed.

When a node serves relays, on the other hand, it stores session
information into the global session cache. If the session's servicer set
is changed, the node needs to update the cached session. Otherwise the
node accepts a claim that should be rejected, or rejects a claim that
should be accepted. As a result, the node ends up with a consensus
failure with `wrong Block.Header.AppHash`.

### Root cause and Fix:

The root cause is simple. We have a call to `ClearSessionCache` in
`EditStakeValidator`, `LegacyForceValidatorUnstake`,
`ForceValidatorUnstake`, and `JailValidator`, but not in
`UnjailValidator`. The proposed fix is to add the call in
`UnjailValidator` as well.