Fix TestValidatorStateChange_EditAndValidateStakeValidator (#1533)
The test `TestValidatorStateChange_EditAndValidateStakeValidator` failed
because `got.OutputAddress` was `nil` but `tt.want.OutputAddress` had a
valid address. Since this test is to test the behavior before NCUST, we
make sure the fetched output address is nil and the rest of fields are
the same as `tt.want`.

The behavior after NCUST is tested through the test
`TestValidatorStateChange_EditAndValidateStakeValidatorAfterNonCustodialUpgrade`,
which passes without failures.