Pass a correct session end height to `NewSession` (#1545)
With session rollover (#1536), we accept relays for older sessions. For
exaple, a node at block 101 accepts a relay for the session height 97.

There is a bug in the relay validation function `relay.Validate`. In the
example above, the function sees the following values:

```
ctx.BlockHeight() = 101
sessionBlockHeight = r.Proof.SessionBlockHeight = 97
sessionCtx.BlockHeight() = 97
```

and if the corresponding session is not cached, it passes `sessionCtx`
and `ctx` to `NewSession`. This may return a wrong session because the
second argument is supposed to be the end of the session, but in this
case it's not.

The proposed fix is if `ctx` is beyond the relay session, we get a new
context of the session end and pass it to `NewSession`.