fix: reviewpad configuration (#1551)
<!-- reviewpad:summarize:start -->
### Summary generated by Reviewpad on 17 May 23 13:34 UTC
This pull request includes a new PULL_REQUEST_TEMPLATE and updates the reviewpad.yml file, removing the workflows section and leaving only the "praise-contributors-on-milestones" workflow.
<!-- reviewpad:summarize:end -->

We understand the the automatic summarization feature was spammy. So now you can use it through markers so that it directly updates the PR description.