Change default config file values (#1568)
## Description

Changes default values for the config file.

<!-- reviewpad:summarize:start -->
### Summary generated by Reviewpad on 05 Jul 23 23:19 UTC
This pull request changes the default values in the `config.go` file.
Specifically, it increases the values for `Mempool.Size` and
`Mempool.CacheSize` from 9000 to 18000. Additionally, it adds two new
fields `Mempool.MaxTxsBytes` and `Mempool.MaxTxBytes` with their
respective values. This patch also updates the `FastSync.Version` to
"v1".
<!-- reviewpad:summarize:end -->