Add a reference to Pocket Pruner (#1572)
## Description

<!-- reviewpad:summarize:start -->
### Summary generated by Reviewpad on 14 Jul 23 15:36 UTC
This pull request adds a reference to Pocket Pruner in the README.md file. It also includes a new document, pruner.md, which provides information about the Pocket Pruner tool and how to use it for servicers and validators. Additionally, it includes contact information for C0D3R support regarding any pruner-related issues.
<!-- reviewpad:summarize:end -->