Part 3 - Correct the condition to write additional parameters
The function `SetParamSet` writes parameters to application db,
where we have a special logic to prevent parameters introduced
after genesis from being written in blocks before those parameters
are introduced.

However, using `IsOnUpgradeHeight` in this logic is wrong because
all post-genesis parameters we have are activated by a feature name,
not when the protocol version is upgraded.  As a result, we had
never exercised this logic.  This didn't cause any real problem yet
because we never call `SetParamSet` for post-genesis parameters
before they're introduced.

We should do the right thing before this bug causes problems in the
future.