Part 2 - MinSignedPerWindow() should return the parameter value
The parameter `pos/MinSignedPerWindow` defines a proportion, but the function
`MinSignedPerWindow` returns the number of blocks, calculated as
`pos/MinSignedPerWindow * pos/SignedBlocksWindow`.  This is confusing.

To resolve this discrepancy, we had "custom logic" to convert the return value
back to the parameter value.  This is unnecessary calculation.

There was indeed a bug in `PosKeeper.GetParams` that returned the return
value from `k.MinSignedPerWindow(ctx)` without the "custom logic".  As a
result, the value of `min_signed_per_window` in the response from
/v1/query/nodeparams was always wrong.

To remove this confusing design, this patch does:
- Rename the function `MinSignedPerWindow` to `MinBlocksSignedPerWindow`
- Add a new `MinSignedPerWindow` function returning the exact parameter value
  in decimal as the other parameter getter functions do