Introduce AppTransfer feature
The patch allows the existing AppMsgStake transaction to change the address of
a staked app, which enables us to *transfer* the existing app slot from one to
a new account without unstaking.

The AppMsgStake message must be signed by the current staked app.  To issue
such a transaction easier, this patch introduces a new `apps transfer` command
that takes the current app's address and a new app's public key.