Upgrade Go version to 1.21 (#1587)
The coming patch #1581 uses Go Generics that is available since Go 1.18.
Since .github/workflows/Dockerfile pulls a container with Go 1.17, that
patch breaks CI workflow. This patch upgrades Go version in all CI
workflows to the latest version 1.21.

## Description

<!-- reviewpad:summarize:start -->
### Summary generated by Reviewpad on 21 Dec 23 08:47 UTC
This pull request upgrades the Go version in all CI workflows to version
1.21. The current version 1.18 is causing issues with the coming patch
that uses Go Generics available since Go 1.18. The upgrade is done in
the .circleci/config.yml file for the build, test, and
trigger-pocket-core-deployments-branches sections. It is also done in
the .github/workflows/Dockerfile file. Additionally, the go.mod and
README.md files are updated to reflect the new Go version. The patch
includes a total of 13 insertions and 10 deletions across 6 files.
<!-- reviewpad:summarize:end -->