Docker Image enhancement and better CLI support. (#1603)
## THIS PR INCLUDES BREAKING CHANGES
- Dockerfile now uses `app` user instead of `root`

#### Why?

A few Geo-Mesh users report to POKTscan about an issue with the new
image after we adopt the one on the pokt-network/pocket-core repository
on the latest RC.

They report that this image is using `root` as the user which is
recommended to be avoided. There are a lot of blogs and documentation
about this,
[here](https://docs.bitnami.com/tutorials/why-non-root-containers-are-important-for-security)
one of them from a well-known docker image user/company.

Also, we detected a few things that could be enhanced on both, entry
point and docker context.

The problem with having a public image using root right now is that
pocket binary generates folders and files that now belong to the `root`
user, so they will need to modify those permissions to belong to the
proper `app` user and group. To this, I added another optional entry
point that could be used once to fix the permission issue and then start
the container as before.


[Here](https://github.com/pokt-network/pocket-core/pull/1603/files#diff-8a9d880ecb3d20b3cfe8def4da1bd200fcbd44131caee95ea699a34faf9cfd6fR19)
you can see how to use it with docker-compose or docker

#### Changes:

- Modifications to Dockerfile allow the container to run as `app` user
instead of `root`.
- Added a new shell script named fix_permissions.sh to fix ownership
issues related to running containers.
- Updated Dockerfile to use this new script.
- Added a .dockerignore file to help maintain a cleaner Docker build
context, excluding unnecessary files.
- Modifications to `entrypoint.sh` allows the user to run all its
internal commands with the proper `--datadir` param. Now properly handle
the start command when `--keybase=false` is sent. Also, allow the user
to pass the `--datadir` as an env variable to omit it on the start
command.