Version bump to RC-0.12.0 (#1615)
Due to a consensus breaking change to fix a regression in RC-0.11.1
(#1614), this patch is upgrading the protocol version to RC-0.12.0. That
regression caused the chainhalt in the testnet at height block. We're
rolling back the testnet to recover it. This version bump helps to make
sure all validators are runing the latest version.

reviewpad:summary